+++ app/Console/Commands/CloneByZonePumpsystems.php
@@ -71,38 +71,40 @@ class CloneByZonePumpsystems extends Command
         try { 
             $farms=Farm::all();
             foreach ($farms as $key => $farm) {
-                $cloningErrors=CloningErrors::where("elements","/farms/id/pumpSystems")->get();
-                if(count($cloningErrors)>0){
-                    foreach ($cloningErrors as $key => $cloningError) {
-                        $pumpSystemsResponse = $this->requestWiseconn('GET',$cloningError->uri);
-                        $pumpSystems=json_decode($pumpSystemsResponse->getBody()->getContents());
-                        $this->info("==========Clonando pendientes por error en peticion (".count($pumpSystems)." elementos)");
-                        foreach ($pumpSystems as $key => $pumpSystem) {
-                            $this->cloneBy($pumpSystem,$farm);
+                if($farm->active_cloning==1){
+                    $cloningErrors=CloningErrors::where("elements","/farms/id/pumpSystems")->get();
+                    if(count($cloningErrors)>0){
+                        foreach ($cloningErrors as $key => $cloningError) {
+                            $pumpSystemsResponse = $this->requestWiseconn('GET',$cloningError->uri);
+                            $pumpSystems=json_decode($pumpSystemsResponse->getBody()->getContents());
+                            $this->info("==========Clonando pendientes por error en peticion (".count($pumpSystems)." elementos)");
+                            foreach ($pumpSystems as $key => $pumpSystem) {
+                                $this->cloneBy($pumpSystem,$farm);
+                            }
+                            $cloningError->delete();
                         }
-                        $cloningError->delete();
-                    }
-                }else{ 
-                    try{
-                        $currentRequestUri='/farms/'.$farm->id_wiseconn.'/pumpSystems';
-                        $currentRequestElement='/farms/id/pumpSystems';
-                        $id_wiseconn=$farm->id_wiseconn;
-                        $pumpSystemsResponse =  $this->requestWiseconn('GET',$currentRequestUri);
-                        $pumpSystems=json_decode($pumpSystemsResponse->getBody()->getContents());
-                        $this->info("==========Clonando nuevos elementos (".count($pumpSystems)." elementos)");
-                        foreach ($pumpSystems as $key => $pumpSystem) {
-                            $this->cloneBy($pumpSystem,$farm);
-                        }
-                    } catch (\Exception $e) {
-                        $this->error("Error:" . $e->getMessage());
-                        $this->error("Linea:" . $e->getLine());
-                        $this->error("currentRequestUri:" . $currentRequestUri);
-                        if(is_null(CloningErrors::where("elements",$currentRequestElement)->where("uri",$currentRequestUri)->where("id_wiseconn",$id_wiseconn)->first())){
-                            $cloningError=new CloningErrors();
-                            $cloningError->elements=$currentRequestElement;
-                            $cloningError->uri=$currentRequestUri;
-                            $cloningError->id_wiseconn=$id_wiseconn;
-                            $cloningError->save();
+                    }else{ 
+                        try{
+                            $currentRequestUri='/farms/'.$farm->id_wiseconn.'/pumpSystems';
+                            $currentRequestElement='/farms/id/pumpSystems';
+                            $id_wiseconn=$farm->id_wiseconn;
+                            $pumpSystemsResponse =  $this->requestWiseconn('GET',$currentRequestUri);
+                            $pumpSystems=json_decode($pumpSystemsResponse->getBody()->getContents());
+                            $this->info("==========Clonando nuevos elementos (".count($pumpSystems)." elementos)");
+                            foreach ($pumpSystems as $key => $pumpSystem) {
+                                $this->cloneBy($pumpSystem,$farm);
+                            }
+                        } catch (\Exception $e) {
+                            $this->error("Error:" . $e->getMessage());
+                            $this->error("Linea:" . $e->getLine());
+                            $this->error("currentRequestUri:" . $currentRequestUri);
+                            if(is_null(CloningErrors::where("elements",$currentRequestElement)->where("uri",$currentRequestUri)->where("id_wiseconn",$id_wiseconn)->first())){
+                                $cloningError=new CloningErrors();
+                                $cloningError->elements=$currentRequestElement;
+                                $cloningError->uri=$currentRequestUri;
+                                $cloningError->id_wiseconn=$id_wiseconn;
+                                $cloningError->save();
+                            }
                         }
                     }
                 }
